Upgrade Castle.Transactions to .NET Standard 2.0
